add expo/cordova compatibility layer
The entry point for the Cordova app is index.js, which uses react-native-web to createRoot on the #appRoot element of index.html

The entry point for an Expo app will be App.tsx (defined in app.json)

As such, a few things need to be moved around – index.js should have nothing except the createRoot business, and App.tsx should contain everything needed for the app including Providers, SafeAreaView wrapper, etc.

To account for Cordova plugins not being available, we can leverage the mocks we have been using for testing; they are basically just minimal JS versions of the plugins (ie using localStorage instead of native DB storage, fetch instead of the ServerCommunication plugin, etc). We'll apply these mocks (in expoCompat.ts) if we detect that we are running in Expo web