more expo compatibility shims
useAppConfig
cordova has a 'deviceready' event that we listen for before we can retrieve the config. Expo does not have this event, so we need to check IS_EXPO and only await if IS_EXPO is false

usePermissionStatus
We will not be needing any permission-enabled functionality on the Expo web version. We can add a usePermissionStatus.web.ts that will automatically be used in place of usePermissionStatus.ts for Expo web, and it has an empty `checkList` so no permissions will be requested.

pushNotifySettings
Similarly, we do not need push notifications on the web version, so we will export a function that does nothing from pushNotifySettings.web.ts