Revert "🔥 Remove the manual gradle installation"
This reverts commit 399e9e7a4f83ed7d67fbf2e249179fa073cfec68.

I _thought_ this wasn't needed because cordova configures `build.gradle` so
that the correct version of gradle is downloaded and installed on the first
run.

However, when I tried to rebuild on a fresh installation, I got the following error

```
Using Android SDK: /Users/kshankar/Library/Android/sdk
Could not find an installed version of Gradle either in Android Studio,
or on your system to install the gradle wrapper. Please include gradle
in your path, or install Android Studio
```

It looks like although cordova installs gradle, the system needs to have a
version of gradle installed as well to bootstrap the gradle wrapper.

This error doesn't show up in the CI/CD pipeline because the CI environment has
gradle installed.

@catarial did you encounter this during your setup as well?! Or maybe it has
not been an issue so far because you are focused on the iOS build.