fix tests, record client stats for missing native consent
in dynamicConfig.test.ts, there are some tests where an error is expected to be thrown and error popup is triggered. clientstats is invoked because we are now recording a ui_error stat on all error popups.
Client stats needs mockDevice and mockGetAppVersion to work

in startprefs.test.ts, a similar error popup occurs so the mocks are needed there too.
However, while inspecting startprefs.ts, I noticied the popup occurs when consent document exists in localstorage but not in native storage.
I do not think we need an error popup for this; instead let's do what we do in storage.ts, which is record a 'missing_keys" client stat, and give it a type of 'document_mismatch'.