simplify mocking by specifying jest setup file
Instead of having each test file call the mocks it uses, we can have jest set up all the mocks beforehand.
in jest.config.js, setupFiles points to setupJestEnv.js where all the mocks are called.

This is way less verbose, and may also speed up the execution of the test suites because the mocks are applied once rather than duplicated for every individual test.

We still have flexibility for individual tests to call mocks on their own; for example in enketoHelper.test.ts there are several calls to mockBEMUserCache(<fakeconfig>) because it is testing out various configs