TimelineContext: only load the last week even if pipeline is behind
This mostly just matters for: 1) testing with historical data, or 2) people whose pipeline is stuck, which indicates a deeper problem.
But it will prevent any scenarios where several weeks/months get loaded in all at once.

To aid with scenario (1), we can add a button to the TimelineScrollList: if there is no travel in the past week and the pipeline end is not in the last week, this gives us a shortcut to the last processed week