MetricsTab: only force >= 14 days on initialization
When the user first goes to the Dashboard in an app session, we want to show them at least 2 weeks so they can compare the previous week to the past week.
But after that we should allow them to select any range.
So we can add an 'isInitialized' bool state to mark this, and only do this ">= 14 days" check on the initialization

Also add a 'refresh' function which not only refreshes the timeline but also resets 'isInitialized' (along with 'aggMetricsIsLoading')