use the new label options and rich modes pattern throughout the UI
Where base modes were previously a property of the label options, and had a separate set of properties, we now have rich modes that inherit from base modes and may include other properties.
This means for the numerous places in the UI where we use mode colors, we should always check the rich mode via get_rich_mode / get_rich_mode_for_value.

Also we can tighten down typings a bit more.
"text" does not exist in label options or in rich modes. It had been added in while label options are being read and the translations are parsed
Instead of that let's just have a function that does this (labelKeyToText). We need to check for translations in 1) the label options being used; and 2) the default label options. if nothing, then convert it to readable (convert underscores to spaces & capitalize)
We'll also have a function that does the opposite.

Updated some typings: LabelOptions type should not accept a parameter. Also should mark REPLACED_MODE as optional

UserInputData should not have "name". This only exists on EnketoUserInputData; it represents the name of the survey