remove StatusBar plugin
I attempted to unify the appearance of the statusbar between platforms by adding the StatusBar plugin (which had already been referenced in the code).
However, this caused major issues on some Android devices with the entire appbar no longer being visible.
This was not a high priority; rather than spend a lot of time trying to get the right appearance, I am just going to revert it for now