👷 Run the tests using npm
Running the test using `npx jest` instead of `npm run test`
means that passing in command line arguments to jest needs to either be coded
in two places, or that jest would be potentially invoked inconsistently.

This changes the invocation to use `npm` so that the command line args (if any)
can be passed in to a single location.
https://github.com/e-mission/e-mission-phone/pull/1193#issuecomment-2558224236