👷 Upload exact versions of the packages used
We use `^` versions for our dependencies, which means that any compatible minor
version could be pulled. However, we recently ran into an issue
(https://github.com/e-mission/e-mission-phone/pull/1193#issuecomment-2558223871)
in which a change to a minor version broke our tests. Sifting through all our
dependencies to find the ones that were bumped, and which dependencies they
bumped in turn is time consuming.

For example in the issue above, the dependency that caused the break was
`react-native`, which we don't include directly in our project.

One option to identify changes would be to check the `package-lock.json`, but
unless we are constantly reinstalling npm, we won't necessarily know what
changed around the time that the tests broke.

Uploading the `package-lock.json` from runs allows us to see the versions from
before and after the break and narrow down the potential changes.