🔥 Stop failing the test if code coverage goes down
We currently run tests and check code coverage in the same workflow,
and fail if the code coverage goes down. But our current coverage is fairly
limited (< 20%) so we have a lot of failures from changes to code that
currently does not have any tests.

Let's still check the code coverage for the patch, but disable it for the
project, so it doesn't mask errors in the tests.

We should not have to enable this since
https://docs.codecov.com/docs/common-recipe-list#set-project-coverage-checks-on-a-pull-request
> By default, Codecov will only show git diff coverage checks on a PR.
> "Project coverage" checks and "project coverage" reporting is not available for Private Repos on the Codecov team plan.
> For all other private repos, and for all public repos, here's how you can also show project coverage checks on a PR.

But it is clearly failing and is the only test failing, so let's make the
change anyway.

Once we improve coverage to cover more areas, we can re-enable this.