fix UI error when enter_fmt_time or exit_fmt_time missing
Confirmed trips always have start_fmt_time and end_fmt_time, but confirmed places are not guaranteed to have enter_fmt_time and exit_fmt_time. Commonly, the last place has not been exited and does not have an exit_fmt_time

Thus, this could be undefined and cause an error when we try to take .substring.
It is ok to

This only affected configs with place inputs (e.g. stage-timeuse) because other configs do not show places in the UI