📌 Check in package-lock.json as well
So we can have the CI use known good versions of the dependencies.
Without this change, the underlying libraries can change based on the loosely
defined semver in the `package.json`, which can cause unexpected failures.

An example of such a failure is:
https://github.com/e-mission/e-mission-docs/issues/1113

Copying over the `package-lock.json` overrides `package.json` and forces the build to use the exact versions specified
https://github.com/e-mission/e-mission-docs/issues/1113#issuecomment-2661545772

Note that `package-lock.json` is apparently designed to checked in to source
control. It used to be checked in (thanks to a contribution from the UNSW team)
but I removed it because every run of `npm install` would override it,
resulting in confusing merge conflicts.

However, it looks like `npm` has now changed its behavior so it respects the
`package-lock.json` if it exists.
https://github.com/e-mission/e-mission-docs/issues/1113#issuecomment-2661550482

So we restore it by copying in known good versions - from
https://github.com/e-mission/e-mission-phone/actions/runs/12448538357
for the "serve" case and from the last successful internal build of NREL
OpenPATH for the `cordovabuild` case.

We also change `bin/configure_xml_and_json.js` to copy these files over as well

Testing done:
- Copying `package-lock.json` into the repo causes `npm install` to work, and
  the tests to pass
- Testing of the setup changes will happen through CI