fix react and react-native versions
react-native 0.78.0 is out now, which uses React 19, but we are still using React 18, causing npm to complain about incompatible versions
Locking down the react-native version to 0.76.7 allows us to stay on React 18 for now.
We do want to go to React 19 eventually, so I am changing React libs to ^18.2.0. This allows 18.3.0 to be used which will give us warnings in the console if we use any deprecated features that won't work in React 19 https://react.dev/blog/2024/04/25/react-19-upgrade-guide#react-18-3

Note that we don't even use the react-native package itself; we just use react-native-web. react-native gets downloaded anyway because it's listed as a peer dependency of @react-navigation/native which we do use.