use reverse-geocoded place names in composite trips
We used to only perform the nominatim queries on the client side, but we added it to the analysis pipeline a while ago. This means processed trips will already have the start and end places' display_name filled in, and it's only for unprocessed trips that we'll need to perform those queries on the client side.
Given this, we can simplify addressNamesHelper considerably. We can just handle the localStorage lookup and/or fetch (if needed) within the useAddressNames hooks rather than the previous approach (which was to explicitly queue all the nominatim requests from LabelTab.tsx and observe as they got put into localStorage)

- extracted GEOCODING_API_URL to separate variable