📍 use reverse-geocoded place names in composite trips (#1205)
* update CompositeTrip and ConfirmPlace types

- Add specific string values for 'key' and 'origin_key' properties – makes it easier to diambiguate types in TS code
- added NominatimResponse type in apiTypes.ts
ConfirmedPlace: add reverse geocode props, fix location type
- getRawEntries return type is a ServerResponse
- Clarified that the CompositeTrip type is AFTER unpacking in timelineHelper. Applied types in timelineHelper; refactored to make clearer

* use reverse-geocoded place names in composite trips

We used to only perform the nominatim queries on the client side, but we added it to the analysis pipeline a while ago. This means processed trips will already have the start and end places' display_name filled in, and it's only for unprocessed trips that we'll need to perform those queries on the client side.
Given this, we can simplify addressNamesHelper considerably. We can just handle the localStorage lookup and/or fetch (if needed) within the useAddressNames hooks rather than the previous approach (which was to explicitly queue all the nominatim requests from LabelTab.tsx and observe as they got put into localStorage)

- extracted GEOCODING_API_URL to separate variable

* add tests for useAddressNames

since the useAddressNames hook is the only export now, let's make it the default export and rename addressNamesHelper -> useAddressNames
Then let's add tests with a couple fake trips/places to make sure it works whether display_name is filled in or not