👤 simplify user profile updates; refactor push notif registration
Before this change, we would make 2 calls to `profile/update` every time the UI is initialized: one for push notification registration and another to store device settings.
We also had a lot of unecessary complexity in all these places (event listeners everywhere!)

All we really need to do on UI init, iff onboarding is done:
1. call `profile/get` on the server to get the current user profile
2. perform push notif registration
3. get current device settings
4. based on the results of the above, determine what fields need to be updated and call `profile/update` on server if needed

1, 2, and 3 can be executed in any order/ in parallel (with Promise.all / Promise.allSettled). Added a file userProfile to handle this.

The added benefit of this is that we have access to the value of the current profile. We can keep it as state in AppContext allowing it to be used anywhere in the UI (this can simplify how custom labels are retrieved + used; and it may be useful for users to know their "last synced ts", for example)
When the profile is updated from the UI, we have to keep the the local state in sync so instead of directly calling commHelper updateUser directly, we expose `updateUserProfile` from AppContext

In removing the complexity, I moved remoteNotifyHandler into pushNotifySettings

Updated tests for all changes