prevent duplicate updateAllUnprocessedInputs call
TimelineContext loadTimelineEntries retrieves the pipeline range, calls updateAllUnprocessedInputs for that range, and then sets the pipeline range as state via setPipelineRange.
Then in LabelTab, we have a useEffect: when pipelineRange changes, we call updateAllUnprocessedInputs (again!)
setPipelineRange is not called from anywhere else, meaning that this duplicate scenario is the time this useEffect gets called. We can remove it.

tested recording new labels, refreshing the page, and changing to different dates, and labels still get matched properly