don't show error popup on reverse geocode failure
This error consistently shows up in the simulator due to SSL not being set up. It could also happen if the API is down or inaccessible for any reason.
I don't want to silence this error altogether, but since it is not a critical failure, I am just going to make it a toast instead of an intrustive popup