refactor Profile page by implementing popup Alerts
The ProfileSettings component has several modals that pop up when a row is clicked, with a lot of repetitive code to track each of these modals' visibility.
It is tidier (and possibily more performant) to instantiate popups as they are triggered. I extended the Alerts mechanism (which previously just handled SnackBar messages / toast messages) to support popups.
We can either pass a title and content, or we can pass a component function that creates a modal

The result is that ProfileSettings.tsx, which was >600 lines and hard to comprehend, is significantly simpler and more approachable

- renamed `AlertManager` to the more concise `Alerts`
- put all the "Developer Zone" components in a subdirectory js/control/devzone, as well as the components for the bluetooth testing page
- extracted a few other modals in ProfileSettings to their own components (NukeDataModal, UploadLogModal, LogoutModal)
- moved providers from index.js to App.tsx and put `AlertBar` inside the providers; this enables popup alerts to use values from contexts, e.g. AppContext
- removed unused / redundant variables and functions throughout